refactor: Move beacon kinds to their respective model package
