+++ common/src/main/java/com/wynntils/functions/LootrunFunctions.java
@@ -7,8 +7,8 @@ package com.wynntils.functions;
 import com.wynntils.core.components.Models;
 import com.wynntils.core.consumers.functions.Function;
 import com.wynntils.core.consumers.functions.arguments.FunctionArguments;
-import com.wynntils.models.beacons.type.LootrunBeaconKind;
 import com.wynntils.models.containers.type.MythicFind;
+import com.wynntils.models.lootrun.beacons.LootrunBeaconKind;
 import com.wynntils.models.lootrun.type.TaskLocation;
 import com.wynntils.utils.EnumUtils;
 import com.wynntils.utils.mc.type.Location;
