+++ common/src/main/java/com/wynntils/models/activities/ActivityModel.java
@@ -16,6 +16,8 @@ import com.wynntils.handlers.scoreboard.ScoreboardPart;
 import com.wynntils.mc.event.ScreenClosedEvent;
 import com.wynntils.mc.event.SetSpawnEvent;
 import com.wynntils.mc.extension.EntityExtension;
+import com.wynntils.models.activities.beacons.ActivityBeaconKind;
+import com.wynntils.models.activities.beacons.ActivityBeaconMarkerKind;
 import com.wynntils.models.activities.caves.CaveInfo;
 import com.wynntils.models.activities.event.ActivityTrackerUpdatedEvent;
 import com.wynntils.models.activities.event.DialogueHistoryReloadedEvent;
@@ -31,8 +33,6 @@ import com.wynntils.models.activities.type.ActivityTrackingState;
 import com.wynntils.models.activities.type.ActivityType;
 import com.wynntils.models.beacons.event.BeaconEvent;
 import com.wynntils.models.beacons.event.BeaconMarkerEvent;
-import com.wynntils.models.beacons.type.ActivityBeaconKind;
-import com.wynntils.models.beacons.type.ActivityBeaconMarkerKind;
 import com.wynntils.models.beacons.type.Beacon;
 import com.wynntils.models.beacons.type.BeaconMarker;
 import com.wynntils.models.character.event.CharacterUpdateEvent;
